+++ lib/template-loader.php
@@ -38,6 +38,10 @@ function get_template_types() {
  * Adds necessary filters to use 'wp_template' posts instead of theme template files.
  */
 function gutenberg_add_template_loader_filters() {
+	if ( ! gutenberg_is_fse_theme() ) {
+		return;
+	}
+
 	foreach ( get_template_types() as $template_type ) {
 		if ( 'embed' === $template_type ) { // Skip 'embed' for now because it is not a regular template type.
 			continue;